server41tests: make renames relative to root of test tree
The pynf tests almost never touch anything outside of the given test
tree, so most work is doable relative to the filehandle of the root of
that tree.  (A rare exception is the --useXXX options which permit the
user to pass absolute paths.)

By redoing lookups each time from the NFSv4 root, we:

	- do longer lookups than necessary;
	- bloat compounds arbitrarily, especially as long as lookups are
	  all done in a single compound; and
	- increase the risk that we'll mistakenly modify something
	  outside of the test tree.

So, we should look up the root filehandle once at initialization, and
perform further lookups relative to that root.

To allow for absolute paths where necessary, introduce the convention
that a path beginning with an empty component is an absolute path.

Start the conversion with rename.

XXX: doesn't work yet.  Probably should introduce absolute paths
separately.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>